Provide an alias for all "default" routes
Alias all "default" routes to enable tests to wait for the respective
responses.  This is in particular helpful if a test verifies
views/behavior that is only present after the respective resource was
resolved. By using aliases instead of using arbitrary timeouts we can
reduce the flakiness of our integration tests.