Fix Ace editor tooltip positioning
Change the fullscreen modal side panel positioning to use position
`right` instead of transform `translateX` as `transform` creates a new
stacking context and the containing block is changing the context which
results in "wrong" positions.

For more details please see:
https://drafts.csswg.org/css-transforms-1/#transform-rendering

With this change in place we no longer need the hacks to correct the Ace
editor tooltip position, as position fixed (used to position the
tooltip) is working as expected.