Fixing regression on error muting
Since the form input fields operate on a different path than the one in the
data, it's not always possible to figure out which error paths to unmute when
the field is edited. This PR introduces a set of rules that defines the input
paths that opt out of the error muting.