Change the JSONEditor state handling
Adjust the JSONEditor state handling to ensure the editor content
properly updates if the internal data is different then the input data,
even when the input data is the same as the initial input data. This
change fixes an issue in which the JSONEditor didn't update if the input
value was the same as the initial value.