refactor(services): remove applyPatch function
This commit will introduce a mayor change in the
way we parse / reduce application configurations
in the service form. Now the ADD_ITEM transaction
will hold null or a complete object of the item instead of
the index. We do this so that we can keep unknown
fields from the JSON. Right now this also is
creating redundancy because all the transactions
after the ADD_ITEM are just reseting the same
value.

E.G. :

The following JSON is parsed:~
```
{
  "array": [{
    "id": "1234",
    "unknownField": "withoutTransaction"
  }]
}
```
And will create the following Transactions:
 - `ADD_ITEM, {id: "1234", unknownField:
   "withoutTransaction", ["array"]}`
 - `SET, ['array', 0, 'id'], "1234"`

But in this case this redundancy is good,in some
places we are depending on the separate
Transactions.

Closes DCOS-14844, DCOS-14756