test(external-volumes): add external volumes test suite
Adds a test case that creates an app with external volume and
tests to see that it is persistent after suspension and resume.
Also, moves the external volume tests from the other test
suites into this newly created one. There is also automatic cleanup
of external volumes via a script that runs during the teardown phase.

Closes DCOS-15073