feat(design-system): add ComponentExample component
Implement ComponentExample component used to display code blocks for design system components. Code can either be in React or HTML. Props can be added to the component to override the generated React and HTML Code. These optional props are: reactCode and htmlCode and are set  with the desired custom content. The component preview is set through the preview prop. It is a mandatory prop unless reactCode and htmlCode are both set.  There are additional props to show/hide these parts: showPreview, showHtml and showReact. These props are set to true by default.

There are several different combinations possible:

    * All (Default with Tabs): preview={preview}
    * Preview Only: preview={preview}, showReact={false}, showHtml={false}
    * Generated React Only: preview={preview}, showHtml={false}
    * Generated HTML Only: preview={preview}, showReact={false}
    * Custom React Only: preview={preview}, reactCode={reactCode}, showHtml={false}
    * Custom HTML Only: preview={preview}, htmlCode={htmlCode}, showReact={false}
    * Custom React: preview={preview}, reactCode={reactCode}
    * Custom HTML: preview={preview}, htmlCode={htmlCode}
    * Custom React and HTML: preview={preview}, reactCode={reactCode}, htmlCode={htmlCode}
    * Omit Preview: showPreview={false}

Closes DCOS-15921