fix(ServiceNetworkingConfigSection): fix system-tests for auto-assign port
In the ServiceNetworkingConfigSection set the port to "Auto Assign" if requiredPorts is false.
If the requirePorts: true then show the selected port.

This fix will fix the system test errors that appear due to this change

Closes DCOS_OSS-1272 (+1 squashed commit)
Squashed commits:
[f875ac5] fix(ServiceNetworkingConfigSection): fix system-tests for auto-assign port

In the ServiceNetworkingConfigSection set the port to "Auto Assign" if requiredPorts is false.
If the requirePorts: true then show the selected port.

This fix will fix the system test errors that appear due to this change

Closes DCOS_OSS-1272