fix(MesosStateStore): not all tasks were returned in a service under a folder
If you create a service inside a folder, it won't show all of the tasks under the service.

In the MesosStateService we had the following check

```
const serviceName = service.getId().split("/").pop();

if(framework.name === serviceName) { ... }

```
with

```
service: { id: "/test/elastic" }
framework { name: "test/elastic" }
```

The problem is that if we have a service in a folder the service id is "/test/elastic" and with the split pop() we lose the full id.
Also the framework's name is "test/elastic". So it would compare framework.name = "test/elastic" with serviceName = "elastic" eventually.

```
  getName() {
    return this.getId().split("/").pop();
  }
```

Now if the service is a framework as default we get the  ``` service.getFrameworkName() ``` and the serviceName as a fallback.

Closes DCOS-17301 (+1 squashed commit)
Squashed commits:
[ecea33d] fix(MesosStateStore): not all tasks were returned in a service under a folder

If you create a service inside a folder, it won't show all of the tasks under the service.

In the MesosStateService we had the following check

```
const serviceName = service.getId().split("/").pop();

if(framework.name === serviceName) { ... }

```
with

```
service: { id: "/test/elastic" }
framework { name: "test/elastic" }
```

The problem is that if we have a service in a folder the service id is "/test/elastic" and with the split pop() we lose the full id.
Also the framework's name is "test/elastic". So it would compare framework.name = "test/elastic" with serviceName = "elastic" eventually.

I created 2 new methods one getFullId() to return "test/elastic" from the service
 and one getServiceName() to return "test/elastic" from the framework.

* Note concerning the naming for "getFullId() and getServiceName()"

I don't like the following method in the Application.js. Although it says getName() it returns part of the Id.
But we can't change it now easily as a defect since it is used in numerous areas.

```
  getName() {
    return this.getId().split("/").pop();
  }
```

Closes DCOS-17301 (+1 squashed commit)
Squashed commits:
[05dbf1d] fix(MesosStateStore): not all tasks were returned in a service under a folder

If you create a service inside a folder, it won't show all of the tasks under the service.

In the MesosStateService we had the following check

```
const serviceName = service.getId().split("/").pop();

if(framework.name === serviceName) { ... }

```
with

```
service: { id: "/test/elastic" }
framework { name: "test/elastic" }
```

The problem is that if we have a service in a folder the service id is "/test/elastic" and with the split pop() we lose the full id.
Also the framework's name is "test/elastic". So it would compare framework.name = "test/elastic" with serviceName = "elastic" eventually.

I created 2 new methods one getFullId() to return "test/elastic" from the service
 and one getServiceName() to return "test/elastic" from the framework.

* Note concerning the naming for "getFullId() and getServiceName()"

I don't like the following method in the Application.js. Although it says getName() it returns part of the Id.
But we can't change it now easily as a defect since it is used in numerous areas.

```
  getName() {
    return this.getId().split("/").pop();
  }
```

Closes DCOS-17301 (+1 squashed commit)
Squashed commits:
[b8c7b71] fix(MesosStateStore): not all tasks were returned in a service under a folder

If you create a service inside a folder, it won't show all of the tasks under the service.

In the MesosStateService we had the following check

```
const serviceName = service.getId().split("/").pop();

if(framework.name === serviceName) { ... }

```
with

```
service: { id: "/test/elastic" }
framework { name: "test/elastic" }
```

The problem is that if we have a service in a folder the service id is "/test/elastic" and with the split pop() we lose the full id.
Also the framework's name is "test/elastic". So it would compare framework.name = "test/elastic" with serviceName = "elastic" eventually.

I created 2 new methods one getFullId() to return "test/elastic" from the service
 and one getServiceName() to return "test/elastic" from the framework.

* Note concerning the naming for "getFullId() and getServiceName()"

I don't like the following method in the Application.js. Although it says getName() it returns part of the Id.
But we can't change it now easily as a defect since it is used in numerous areas.

```
  getName() {
    return this.getId().split("/").pop();
  }
```

Closes DCOS-17301