fix(JobsList): add column sorting for Status and Last Run
Previously, jobs were only able to be sorted by Job Name.
This change allows sorting by Status and Last Run. Both these columns
are sorted in order of "importance to the avg user". For example,
jobs that are failing are higher importance than successful jobs,
and will be sorted to the top of the list.

Closes DCOS-16744.