refactor(DOMUtils): remove isElementOnTop util
Remove the `isElementOnTop` DOM util as the util is only used to check
whether the `Chart` component should be updated.
The checks performed by the utility force layout/reflow (for a
comprehensive list of properties or methods that force layout/reflow
see:  https://gist.github.com/paulirish/5d52fb081b3570c81e3a) and as
such shouldn't be used to check whether a component should be updated.
The util also always returns true if an element is off the screen which
completely renders the utility useless if used as a should component
update condition.