fix(scripts): adjust Marathon RAML install the version
Adjust the Marathon RAML URL pattern and update the version to fetch the
latest Marathon 1.5 API Spec and ensure that the UI is compatible with
Marathon.
The latest spec includes `unreachableStrategy` adjustments that if not
applied prevents users from creating/editings apps that are configured
to expunge immediately or flag apps as inactive.

Closes DCOS_OSS-1582