fix(ServiceDeleteActions): remove promises from service delete actions
Cleanup for the Service Delete Actions. Deleted the promises and the deleteGroup where we created a promise array of all of the delete actions for the group's frameworks.

Closes DCOS-18006 (+1 squashed commit)
Squashed commits:
[5433a81] feat(MarathonActions): uninstall SDK services like standard services

Until now when the user tried to delete a Framework or SDK Service from the UI we displayed a message that was referencing the CLI commands.

With this PR, we can delete frameworks and SDK Services from the UI.

When we delete a group the standard services are correctly deleted as well. However, Cosmos and Marathon don't communicate with each other, so if we have Framework/ SDK Services in the group these aren't correctly uninstalled.
Solution: We show a popup that lists the frameworks running in the group so the user will have to first uninstall them separately before deleting the group.

Closes DCOS-18006