feat(ConfigurationMapAction): add href prop
Add an href prop to this component. When the href prop
is provided it will render an anchor tag. When the href
prop is not provided it will render a button tag with
the onClick prop provided. This is more semantically correct
HTML: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#Accessibility_recommendations

Closes DCOS-19339.