refactor(Mount): allow array of children for mount service
This commit changes the MountService so that it accepts multiple children
as default value, it also adds a test to verify the correct rendering
of multiple children.

This allows us to pass an array of children to the Mountservice.
`Mount.render` wraps multiple elements in a div by itself, so it
is able to receive multiple children.
This commit changes the `PropTypes.children` definition to reflect
this.