fix(test): add flags to identify and stub Mesos requests
Add flags (query parameters) to identify and stub Mesos Operator API
requests to fix the integration tests that were otherwise failing or
flaky due to parsing errors in the `MesosStateStore`. The store failed
to parse the `GET_MASTER` response as we could only register one stub
for both the `GET_MASTER` and `SUBSCRIBE` requests resulting in JSON
parse errors due to the different response formats (RecoredIO<JSON>,
JSON).

The flags are necessary as Cypress can only stub requests based on the
URL and request method and doesn't provide the means to dynamically stub
requests based on the payload as documented in
https://github.com/cypress-io/cypress/issues/521. We will remove the
flags once the Cypress team rewrote the network handling and added
support for dynamic stubs as documented in
https://github.com/cypress-io/cypress/issues/687 or we've  addressed
https://jira.mesosphere.com/browse/DCOS_OSS-2021.