fix(FullScreenModal): remove animation and fix loader
This fixes the flashes of dark-grey when opening and closing
full screen modals. This is a two part fix. Part one is making the
loader background white. This because the background of DC/OS is now
dark-grey, so now we must manually set the loader background to white.
Part two if this change is removing animations from the full screen
modal. This is smoother, and prevents short flashing of the dark-grey
background.

Closes DCOS-20343