refactor(Deployment): upload release to S3 after build
This commit creates a tarball and uploads it to the correct
S3 bucket.
The original plan was to do this after the tests ran successfully in the main
jenkinsfile, but the tests are not green there.
This is the reason why we thought it might make sense to split the release in a
seperate Jenkinsfile instead as a first step.

Closes DCOS-20468