fix(ServiceForm): fixes focus/change/blur behaviour
This commit changes the onChange handler of the ServiceForm so that
it no longer implicitly sets state.editingFieldPath, instead a new
internal function `handleFormFocus` is introduced.

The old implementation was broken and also was one of the reasons
why our tests were flaky, in fact some tests had to be fixed
after changing the form - they were expecting the broken state.

Also the previous implementation caused the form to lose information
when a radio or checkbox inpout was changed last in some situations.

Firefox for Example doesnt focus a radio input and therefore
never fires a blur event.

Closes DCOS-21122