fix(ServicesContainer): checks and updates modalProps if needed
This commit adds a `getCorrectedModalProps` method to the `ServicesContainer`
which cleans and updates the `modalProps` object which is passed down to
the modals.
Before this change it was possible to have an invalid state in `modalProps`
like this: { service: undefined, id: "delete" }. This should be fixed now so
that `service` is always set, if `id` is. If anything is wrong the method
tries to correct the object as good as possible.

Closes DCOS-21359