feat(EditServiceModal): show cosmos services in json-schema form
Show not only Framework's in the json-schema form, but all
services installed with cosmos. We detect this with two labels
DCOS_PACKAGE_DEFINITION and DCOS_PACKAGE_METADATA. Frameworks are
a subset of all services with these labels. This change also
shows the cosmos/servic/descibe configuration in the configruation
tab.

Closes DCOS-21614