fix(DSLInputField): adjust change event handling
Remove the change event debouncing to resolve issues in which we lost the filter expression data.

The component was designed to be _stateless_ meaning that the hosting component is responsible for updating the `expression` property. Debouncing the event potentially leads to data loss as the hosting component might overwrite the expression before the change event was fired.

Closes DCOS-22215