feat(JobDetailPage): this connects the job detail page to the data layer
This introduces the data layer to the Job Detail Page. This removes most of the usage of metronome
store. Most data now comes from the datalayer and uses a mediator. Also the configuration and
histroy run page use the new data format instead of the old structs.

Close DCOS-37615 DCOS-37619