fix(MesosStream): stop replay multicasting the stream
Adjust `MesosStream` to stop multicasting the stream using the replay subject. This change allows the browser to garbage collect dependent/sliced strings created while processing the Mesos response.

Please note that this is a temporary adjustment as it would result in multiple open connections if consumed by more than one component. We're going to revert this change once the `http-service`  to emit event objects instead of the response as well as changed the `mesos-client` to output parsed JSON.

Closes DCOS-38339