feat(intercom): add intercom plugin
Adds intercom plugin: appends intercom code to the DOM.
This enables the intercom widget in the bottom right corner
of the page. The widget appears on login and disaappears on
logout.

Leverages IntercomStore to store intercom attributes. Store
listens to various stores (MetadataStore, NodeHealthStore,
ConfigStore) and updates when a change event occurs. It will
in turn notify Intercom to update with its attributes.

Closes DCOS-40519