feat(FrameworkConfiguration): validate on submit instead of every change
I introduced some damage here due to the way that react-jsonschema-form handles validation errors.

I had to reach into the internal state of the SchemaForm after submitting the form due to a limitation with react-jsonschema-form but this might not be a big deal.

liveValidate is not desirable until the form is submitted. When liveValidate is false, the onError prop is called when the form is submitted with a flattened list of error messages only suitable for an anchored list. Conversely, the entire state
is exposed to "onChange". However, "onChange" does not contain errorSchema until the form is submitted (and validated).

By storing a ref to SchemaForm and accessing its state, I'm effectively
giving as much information to "onError" as is given to "onChange"

Calls to onChange given entire state:
https://github.com/mozilla-services/react-jsonschema-form/blob/v1.0.4/src/components/Form.js#L118

Calls to onError given error array
https://github.com/mozilla-services/react-jsonschema-form/blob/v1.0.4/src/components/Form.js#L143

Validation only invoked for onChange when liveValidation enabled:
https://github.com/mozilla-services/react-jsonschema-form/blob/v1.0.4/src/components/Form.js#L106-L107

(Links are for v1.0.4, but this note was also true for our current version 0.50.0)