feat(StringUtil): filterByString now sorts by search relevance
Relevance scoring is simplified and was inspired by Sifter.js.

In practice, each object uses a set of properties that extract a string body which is analyzed against a search phrase. The phrase is tokenized, with each token searched for within the body. The scoring function accounts for a combination of phrase length and proximity to the beginning of the search text. The highest scoring token is used for each object.

Example:

Searching "kafka" would score these texts in the order C, B, A since kafka appears closest to the beginning of the string in item C.

A: "confluent-kafka"
B: "beta-kafka"
C: "kafka"

Test expectations have been modified to account for being sorted by relevance.

Closes DCOS-21232