fix(PackageDetailTab): ensure package name is rendered
The package name was not being properly rendered when translated using
the `<Trans>` macro. After trying several different changes without any
luck, I decided to move the whole message to a variable and remove the
Trans macro. This will ensure we translate this message later when we
address translating strings from variables.