refactor(ClipboardTrigger): Remove optional props for tooltip text
Only the default props were used, so I marked those for translation.

Also fixes a small bug with copyText, which allowed react nodes but should only allow strings since that is the value to be copied to the clipboard. Callers were always passing in strings, but no warning was given because strings are valid nodes.