fix(FrameworkConfigurationForm): writeErrors schema
When attempting to `writeErrors` we check that the `formData` is an object,
but if so we assume the schema has properties. But if the `formData` is an
Array then its an object and the schema does not have any properties.

Added a check that the schema has properties before iterating over
`formData` keys. To resolve this issue.