feat: SDK Service Plans data-layer
Introducing data-layer for the services plugin. We are adding this to
support querying for a SDK service's plans. These will be displayed in
the UI. To support querying other service information through the
data-layer the root query object is the `Service` even though currently
this only supports a name and the plans.

- Closes DCOS-43910