feat(types): enable type checking
we can see type errors in the console now. also the command triggered to build
for production returns an error code if a type error has been found. need to
check whether this is sufficient to break on CI.

`noEmit` is added to stop tsc from generating .js-files when running tsc without

the --noEmit flag.
Closes DCOS-48252