feat(ui-update): show toasts when UI Updated
Adding stream subscriptions to ui-update plugin that will push Toast
notifications when the UI is updated, an update fails or a rollback fails.
To support this I abstracted the streams from the `UIDetails` component
to the new `streams.ts` file. This allows the streams to be shared
between the component and notifications.

Closes DCOS-45132