chore(jest): express our preprocessor more jesty
introduce babel-plugin-module-resolver to solve module-aliases on a
transpiler-level instead of having separate webpack- and jest-configs.

remove then unneeded dependencies.

remove the resolution of .es6-files in jest. it's been there forever and seems
to not be needed anymore.

use mock-files (like suggested in the jest-docs) instead of custom
preprocessor-rules for images/files.

introduce an isolated and minimal jison-preprocessor.