feat(ServiceTree): aggregate status values
Introduced the `ServiceTreeStatusSummary` response type that is returned
from `ServiceTree.getStatus()` if the tree contains more than 1 status
value to aggregate. This summary object can be used to render a status
that can also be translated using the id & value key values. It also
contains the highest priority statuses display name and a count of
statuses grouped by the StatusIcon.name. The highest priority status
count is used for the displayed status values, but all are returned so
they can be used to generate the desired tooltip content that will be
implemented in another ticket.

- Closes DCOS-51000