refactor: add sorting-attr to service statuses
get rid of FrameworkStatus again as it seems more futureProof to not make a
special case for frameworks here. also a sorting attribute has been introduced,
while the key has been removed - it was sometimes (seemingly falsely) used for
sorting and sometimes to identify a particular status by matching on the
key-attribute. as statuses are constants, we can match those directly, making
the key obsolete.