chore: remove StatusMapping
it looks like this has been introduced 2016 to differently highlight link to
services that are running. we don't seem to have that concept anymore and thus
don't need the class nor the mapping that was duplicated in some place and
creates a lot of noise in generating that very class dependent of the the
statuses displayName ("Running").

We deleted the the last occurence of the `subTitleClassName`-prop of our
DetailViewHeader-component which has been deleted accordingly.