chore: make use of exhaustiveness checking
removing the default-returns in favour of fully exhausted switch-cases.

this makes typescript tell you to implement a mapping, if you add another status.

it may look brittler than having a value in every case that "saves" our backs.
usually the default value is unwanted, if we tried all cases we anticipated for
but did not find one. in this case i personally think it's better to let the app
blow up and fix the root cause.