fix: remove deploying service filter test
Removed the "Filters Deploying services" integration test. With the new
feature of aggregate service statuses for the group based on the status
category the status text for deploying group vs service differs.
"Processing" vs "Deploying" this is intentional and there is not a great
way to fix this test without further changes the UX. At least until we
improve filters or change the category copy this test would fail.