refactor(ServicesTable): remove Util, inline sorting
sorting feels very much like the responsibility of the table and the indirection
that's been happening before revealed that the sorting for regions is not even
implemented.
now that we don't have the double indirection (getting sorter from Column-file,
determining sorter in Util) a mismatch should be more obvious.
also we had sorting for `tasks` before. that's been removed, because i did not
find an according tasks-column.

the tests for sorting have been rewritten, because i was not clever enough to
understand the old ones (i made sure that the data stood the same though).

also a Version-Util has been factored out, as it seemed natural to have some
place to stuff the version-things to outside of the table.