fix: form-colons
factored out separate header-sections that are independent of the lines to
follow after. this way we have a definitive border to measure against: the top
of a form row. note that the bottom of a row is not suitable as it can vary
depending on validation states (see jobs run config with duplicate keys).

this has been done in preparation of the new "environment" tab in the jobs form
that looks very broken if this issue in not addressed.

Closes DCOS-45215