refactor: isOnlyWhitespace
use the conditional as return directly.

i tried to refactor `${str}` to be just `str`, but it appears that despite the
signature saying that only string will come in other stuff comes in and breaks
the code. thus the signature has been changed to clarify that this function not
only intents to check strings.