fix(ui-update): parse build version in data-layer
Updated the ui-update data-layer to parse the build version from string
provided instead of relying on semver.coerce. We currently use the
template of `<branch>+<build_version>(+<build_hash>)?` when branch is
`master` coercing the value gives use the build version, but if the
build is made from a release branch, `1.13` then `1.13` is returned from
semver.coerce instead of our desired build verions after the first `+`.

I added a `parseVersion` utility to the ui-update data-layer to handle
getting the build version from the string assuming the template is
followed. Unit tests for the utility and data-layer were added as well.

- Closes DCOS-53297
- Closes DCOS-53295