chore: save on libraries
lodash's memoize did not behave as intended in the first place because it only
looks at the first argument of the function to memoize to determine a cache key.
by providing a "resolver"-function (a fancy term for a function that determines
a custom cache key) we can get around this limitation without introducing a new
library.

also during the review of https://github.com/dcos/dcos-ui/pull/3860 it became
apparent that we are about to remove the last imports stating
`import _ from "lodash"`, which means that we'll save on bundle size be touching
the NodesTable as well! See the linkes PR for screenshots of the bundle
analzyer. In short: we are saving a 130k (not gzipped of course).