feat: allow to route packages without a version
this aims to solve the problem what let to
https://jira.mesosphere.com/browse/DCOS-52749 without sticking to the solution
proposed by that very issue.

instead of adding a redirect, we make use of the cosmos-api always responding
with data for `latest` if no version was given. as we previously assumed that a
version had to be specified always, that assumption has been factored out, only
sending a version-query-param if it is actually set. this evades the need for a
redirect altogether and seems like a more flexible solution.

also someone updated prettier without running it over the whole codebase, so
a dot has to be moved to comply with our linting...