feat: check if a package has a dep before allowing to install
some packages (currently only kubernetes-cluster) have a `manager` that must be
installed before they can be used. currently the API restricts the installation
of kubernetes-cluster if kubernetes is not installed yet. you have to go through
the whole installation-process in the UI to afterwards be notified that stuff
did not work. as a remedy a check has been implemented that renders an infobox
with a nice warning as well as the reviewAndRun-button disabled - so now we
can't waste a lot of work in the first place when we don't know about
dependencies upfront.