chore: add crypess:open to make better use of cypress during development
if you open cypress and change a test file, it actually reruns that file. when
paired with `.only` on a single test case, it looks like writing integration
tests will be waaaay more convenient for me, as the feedback cycle actually
feels quite convenient.

also make some other commands use the version that we specify in the
package.json.