refactor: rename JobHistoryRun to HistoricJobRun
that name feels easier to read for me throughout the code.

the optional tasks-property has been introduced, as metronome now ships that
data for all runs from now.

the disciminator for historic vs active JobRuns has been changed from `jobId` to
`finishedAt` as it feels more natural (i checked with the metronome code that
this is a valid option)