feat: support `enforceRole` in ServiceTree
Added support for the forthcoming `enforceRole` property for groups from
Marathon. Not all groups are expected to have the property so this new
property and getter function can be undefined.

Added a `getGroups` function as well to support getting the groups from
a ServiceTree like we support `getServices`.

Added handling for `enforceRole` to `MarathonUtil.parseGroups` since
this is where the fetch response is processed initially.