feat: groups data-layer with quota
Implement Services data-layer extension with groups query for returning
quota data related to a group.

The group data is retrieved from the MarathonStore since these queries
will be run under the services container and its using the MarathonStore
to poll for groups. We should research refactoring this to invert the
dependency and instead have the MarathonStore get groups from the
data-layer and the data-layer make the request to the API. But this was
a larger refactor then I wanted to include in this work.

- Closes DCOS-55630