refactor(NodesTable): simplify sorting
by narrowing the interface of `updateData` and only returning the sorted data,
we now save some boilerplate.

also the logic to find out whether the `sortColumn` or `sortDirection` changed
is condensed by only making structural changes to the code.

at first i made sure to maintain the optimization that evades the need to sort
again if the sortDirection or sortColumn have not changed. then i figured that
this is a bug as new entries won't be properly sorted until the user sorts
something again.