refactor(DSLFilters): convert a first filter to typescript
remaining backwards compatible. everything should still work.

to properly operate on all the types involved, the `List` was made generic and
its types narrowed down to something sensible.

Also some complexity around the NodesRegionFilter has been removed. It
previously took a list of regions to filter for as an argument. That list would
always be the complete Set of the regions present in the NodesTable thus not
limiting the result set in any way.