refactor(DSLFilters): remove DSLFilterList
DSLParserUtil was the only consumer of DSLFilterList.
By inlining its functionality (`getMatchingFilters`) we can safely remove it
by making a single line change:

```diff
-      .getMatchingFilters(ast.filterType, ast.filterParams)
+      .filter(f => f.filterCanHandle(ast.filterType, ast.filterParams))
```

The trickery in the constructor of DSLFilterList is remarkable. If you have a
two spare minutes go check it out!

Admittedly there are a couple of Prop-checks that are less strict now - i don't
think that this would have prevented errors though as the console is noisy
anyways...