refactor: groups data-layer
Moved Quota limit string and role counts into the quota type for the
graphql resolver. Also updated the counts to use a single reduceItems
instead of building them separately and iterating over the ServiceTree
twice.

This is still a less then ideal solution since we build the role counts
before we know if the group has quota limits. And we could skip those
counts if the group has no quota. But that is something we could try to
resolve when we switch from filtering on enforceRole to has quota limits.

I'm also not entirely happy with the type key names I added here.