chore(QuotaOverviewLimitColumn): normalize icons
Ensured the icons all used the same parent div and classes. "Partially
Enforced" was the only one using the parent div since it was the only
one with a tooltip, but this caused its icon to be slightly wider then
the others and the status text was not aligned because of this.