fix(NodesTable): show latest data from state summary
there has been an effort to optimize the performance of the nodesTable.
unfortunately one of the optimizations turns out to not be applicable:

when only calling `setFilters` when filterData has changed within
`componentWillReceiveProps`, we're effectively ignoring the data from the
summary-endpoint, which contains data regarding the number of running tasks,
drain_info or resource consumption.

a change of this data needs to trigger a rerender. we're now assuming that
the underlying data may change whenever new props arrive. we may have
opportunities to be smarter here, but for now it's better to have changes
reflecting while the table is open than only rerendering when someone uses
sorting or filtering (or a health-value changing - although i did not confirm
that this worked before).

it looks like in the past 3 month nobody noticed that the table is not
displaying realtime data. maybe that's worth a discussion or at least a
separate set of integration-tests.

Closes DCOS-56382