refactor: groups data-layer
Removed the filter argument from the groups query in favor of filters on
the result stream for groups that have quota.

I introduced `QuotaUtil` and moved the new has QuotaLimit functions,
existing getQuotaLimit function and extracted code from the quota
resolver to the populateResourcesFromRole function. I did this to groups
quota related functions and easily unit test all of these functions.

I created `groupHasQuotaLimit` and `quotaHasLimit` as separate functions
so that the same logic could be used by the Overview component and the
resolver even though they were operating on related but different types.

- Closes DCOS-56352