refactor(QuotaUtil): moved limit status strings to constant
Moved the Quota Limit statuses to a constant object and removed type to
ensure lingui doesn't lose the ids. There may be a better way to resolve
this issue and I'm happy to adjust this as we lost the benefit of the
type in the switch case in the column renderer.